Add more comment in MaxSizePalindromeCount

+++ src/Palindrome.h
@@ -21,5 +21,5 @@ public:
     size_t PalindromeAnagramCount1(string const &);
     void HighestValuePalindrome(string &, size_t);
     size_t ShortPalindrome(const string &);
-    long double MaxSizePalindromeCount(string const &, size_t, size_t);
+    long MaxSizePalindromeCount(string const &, size_t, size_t);
 };
